// packages/lib/composer/ogcsta2chart/src/interfaces/OGCSTAToChartData.ts
/*********************************************************************
 * Copyright (c) 2025 Contributors to the Eclipse Foundation.
 *
 * This program and the accompanying materials are made
 * available under the terms of the Eclipse Public License 2.0
 * which is available at https://www.eclipse.org/legal/epl-2.0/
 *
 * SPDX-License-Identifier: EPL-2.0
 *
 * Contributors:
 *   Smart City Jena
 **********************************************************************/

export interface DatastreamSelection {
  datastreamId: string
  label?: string
  color?: string
}

export interface OGCSTAToChartData {
  labels: string[]
  datasets: {
    label: string
    data: (number | null)[]
    backgroundColor?: string
    borderColor?: string
  }[]
}